fixed min-maxPrice issue altering Float with Double for min and maxPrice. Also updated tests to keep up with the changes

+++ src/main/java/com/petros/bookstore/controller/BookController.java
@@ -17,7 +17,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.IllegalFormatCodePointException;
 
 /**
  * REST controller for managing books in the bookstore.
@@ -60,8 +59,8 @@ public class BookController {
             @RequestParam (required = false) String author,
             @RequestParam (required = false) @Min(0) Integer availability,
             @RequestParam (required = false) Genre genre,
-            @RequestParam (required = false) @DecimalMin("0.0") @Digits(integer = 5, fraction = 2, message = "decimals up to 2 digits") Float minPrice,
-            @RequestParam (required = false) @DecimalMin("0.0") @Digits(integer = 5, fraction = 2, message = "decimals up to 2 digits") Float maxPrice,
+            @RequestParam (required = false) @DecimalMin("0.0") @Digits(integer = 5, fraction = 2, message = "decimals up to 2 digits") Double minPrice,
+            @RequestParam (required = false) @DecimalMin("0.0") @Digits(integer = 5, fraction = 2, message = "decimals up to 2 digits") Double maxPrice,
             Pageable pageable) {
 
         if ((minPrice == null) != (maxPrice == null)) {
@@ -72,12 +71,6 @@ public class BookController {
             throw new IllegalArgumentException("minPrice cannot be greater than maxPrice");
         }
 
-//        if (minPrice != null && maxPrice != null){
-//            minPrice = (float) (Math.floor(minPrice * 100) / 100.0);
-//            maxPrice = (float) (Math.floor(maxPrice * 100) / 100.0);
-//            System.out.println(minPrice);
-//            System.out.println(maxPrice);
-//        }
 
         if (title != null || author != null || availability != null || genre != null || (minPrice != null & maxPrice != null)) {
             return bookService.searchBooks(title, author, availability, genre, minPrice, maxPrice, pageable);
